+++ pkg/tracing/provider/stackdriver/tracer.go
@@ -1,4 +1,4 @@
-package gcloud
+package stackdriver
 
 import (
 	"context"
